# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClearLedger accounting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/journal-entries:
    get:
      operationId: listJournalEntries
      tags: [journalEntries]
      summary: List journal entries
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journalEntries]
      summary: Create a manual journal entry
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # Reconciliations
...
  /businesses/{businessId}/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /businesses/{businessId}/dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Get 12-month revenue vs expenses chart data
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyChartPoint"
  /businesses/{businessId}/dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions for dashboard
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
        accountId:
          type: ["integer", "null"]
        accountName:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, plaid, upload]
        reconciled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - businessId
        - date
        - description
        - amount
        - type
        - source
        - reconciled
        - createdAt
...
    JournalLine:
      type: object
      properties:
        id:
          type: integer
        journalEntryId:
          type: integer
        accountId:
          type: integer
        accountName:
          type: ["string", "null"]
        debitAmount:
          type: number
        creditAmount:
          type: number
      required:
        - id
        - journalEntryId
        - accountId
        - debitAmount
        - creditAmount
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        date:
          type: string
        memo:
          type: ["string", "null"]
        createdAt:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/JournalLine"
      required:
        - id
        - businessId
        - date
        - createdAt
        - lines
    CreateJournalLineInput:
      type: object
      properties:
        accountId:
          type: integer
        debitAmount:
          type: number
        creditAmount:
          type: number
      required:
        - accountId
        - debitAmount
        - creditAmount
    CreateJournalEntryBody:
      type: object
      properties:
        date:
          type: string
        memo:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CreateJournalLineInput"
      required:
        - date
        - lines

    # Reconciliation schemas
...
    DashboardSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        cashBalance:
          type: number
        reconciliationStatus:
          type: string
          enum: [up_to_date, needs_review]
        transactionCount:
          type: integer
      required:
        - totalRevenue
        - totalExpenses
        - netProfit
        - cashBalance
        - reconciliationStatus
        - transactionCount
    MonthlyChartPoint:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
      required:
        - month
        - revenue
        - expenses