# lib/api-zod/src/generated/types/createJournalLineInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearLedger accounting API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJournalLineInput {
  accountId: number;
  debitAmount: number;
  creditAmount: number;
}