# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearLedger accounting API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryReconciliationStatus } from "./dashboardSummaryReconciliationStatus";

export interface DashboardSummary {
  totalRevenue: number;
  totalExpenses: number;
  netProfit: number;
  cashBalance: number;
  reconciliationStatus: DashboardSummaryReconciliationStatus;
  transactionCount: number;
}